tests/components/ui/Badge/index.test.tsx
import { render, screen } from "@testing-library/react";
import { Badge } from "@/components/ui/Badge";

describe("Badge Component", () => {
  it("renders correctly with default props", () => {
    render(<Badge onClick={() => {}}>Click Me</Badge>);
    const element = screen.getByText(/Click Me/i);
    expect(element).toBeInTheDocument();
  });

  it("applies the correct variant class", () => {
    render(
      <Badge variant="destructive" onClick={() => {}}>
        Delete
      </Badge>
    );
    const element = screen.getByText(/Delete/i);
    expect(element).toHaveClass("variant-destructive");
  });

  it("applies the correct size class", () => {
    render(
      <Badge variant="destructive" size="lg" onClick={() => {}}>
        Delete
      </Badge>
    );
    const element = screen.getByText(/Delete/i);
    expect(element).toHaveClass("size-lg");
  });
});